Load channels on app init.
