+++ src/store/reducers/channels/index.js
@@ -1,10 +1,12 @@
+export const SET_CHANNELS = 'SET_CHANNELS';
+
 const initialState = {
     items: [],
 }
 
 export default function(state = initialState, action) {
     switch(action.type) {
-        case 'DO_THIS':
+        case SET_CHANNELS:
             return {...state, items: action.payload.channels};
 
         default:
